+++ R/setInitialValues.R
@@ -47,7 +47,7 @@ setInitialValues <- function(params, sim) {
     }
     params@initial_n[] <- sim@n[no_t, , ]
     params@initial_n_pp[] <- sim@n_pp[no_t, ]
-    params@initial_n_other <- sim@n_other[no_t, ]
+    params@initial_n_other[] <- sim@n_other[no_t, ]
     params@initial_effort <- sim@effort[no_t, ]
     params
 }