+++ tests/testthat/test-setInitialValues.R
@@ -24,7 +24,17 @@ test_that("We can set and get initial values from sim object", {
 test_that("Can set initial values in a model with a single species", {
     species_params <- NS_species_params[1, ]
     params <- newMultispeciesParams(species_params)
-    sim <- project(params, t_max=0.1)
+    sim <- project(params, t_max = 0.1, t_save = 0.1)
     p <- setInitialValues(params, sim)
-    expect_identical(finalN(sim), initialN(params))
+    expect_identical(finalN(sim), initialN(p))
+})
+
+test_that("Can set initial values in a model with a single other component", {
+    params <- setComponent(params, 
+                           component = "test",
+                           initial_value = 1,
+                           dynamics_fun = "test_dyn")
+    sim <- project(params, t_max = 0.1, t_save = 0.1)
+    p <- setInitialValues(params, sim)
+    expect_identical(initialNOther(p), list(test = 111))
 })